requirement.txt
torch==2.11
torchvision==0.16.1
torchsummary==1.5.1
matplotlib==3.8.2
scikit-learn==1.3.2
scikit-image==0.22.0
Pillow==10.1.0
pandas==2.1.4
numpy==1.26.2
seaborn==0.13.0
scipy==1.11.4